+++ src/main/java/com/digitalnogran/email/DigitalnogranApplication.java
@@ -1,4 +1,4 @@
-package com.digital.nogran.api;
+package com.digitalnogran.email;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
