+++ src/main/java/com/digitalnogran/email/repository/ContatoRepository.java
@@ -1,8 +1,8 @@
-package com.digital.nogran.api.repository;
+package com.digitalnogran.email.repository;
 
 import java.util.List;
 
-import com.digital.nogran.api.model.Contato;
+import com.digitalnogran.email.model.Contato;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
